refactor: extract shared compound-editor state factory (#762)
Lift the inline editor logic out of compound-editor-modal.svelte (1,258 LOC)
into a rune-based state factory so the desktop modal and the mobile sheet are
thin views over one controller. Structure-only, no behaviour change.

- NEW state/compound-editor.svelte.ts: createCompoundEditorState owns
  composition rows, atom-count vs %-by-mass mode, duplicate-Z detection, sum
  tracking / rescale-to-100, formula & preset application, deferred validation,
  URL-recovery highlighting, and save/cancel. Pure helpers in
  utils/compound-derive.ts and utils/formula-parser.ts stay the callees.
- NEW compound-editor/desktop-sheet.svelte: desktop view bound to the
  controller (picker + preset/formula confirm dialogs live here).
- compound-editor-modal.svelte reduced to a 111-LOC orchestrator that creates
  the controller, runs the open/load() lifecycle, and picks desktop vs mobile.
- mobile-sheet already bound to the shared EditorController, so no logic is
  duplicated across views.
- Direct factory unit tests; existing component + e2e tests unchanged.
- docs/03-architecture.md §8 + listings updated.
