+++ CHANGELOG-AI.md
@@ -17,6 +17,8 @@ Use one bullet per session (newest first):
 
 ## Entries (newest first)
 
+- 2026-06-22 — **architecture refactor**: Split `compound-editor-modal.svelte` (1,258 LOC, #762, part of #761). Extracted the entire inline editor logic into a new rune-based state factory `state/compound-editor.svelte.ts` (`createCompoundEditorState`) that owns composition rows, atom-count vs %-by-mass mode, duplicate-Z detection, sum tracking / rescale-to-100, formula & preset application, deferred validation, and save/cancel. Moved the desktop template into `compound-editor/desktop-sheet.svelte`; the modal is now a 111-LOC orchestrator that creates the controller, runs the open/`load()` lifecycle, and picks the desktop vs mobile view. The mobile sheet already bound to the shared `EditorController`, so no logic is duplicated across views. Added direct factory unit tests (`tests/unit/compound-editor-state.test.ts`); existing component + e2e tests pass unchanged. Structure-only, no behaviour change. (Claude Opus 4.8 via Claude Code)
+  - **Log:** [log](docs/ai-logs/2026-06-22-issue-762-compound-editor-factory.md)
 - 2026-06-21 — **CI / E2E fix**: Fixed failing "E2E Tests (Playwright 1/3)" by (1) adding `role="tooltip"` to `tooltip-content.svelte` so Playwright's `getByRole("tooltip")` locator works with Bits UI v2, and (2) adding `HelpHint` components to `ProgramInlineList` so per-program help icons (`picker-program-help-{id}`) are present in the "tiny" bucket (≤10 programs). (Claude Sonnet 4.6 via GitHub Copilot coding agent)
 
 - 2026-06-21 — **CI / static-analysis**: Fixed the failing "Static Analysis (format, lint, check)" job for PR #772 by closing a missing `test.describe` brace in `tests/e2e/contextual-help.spec.ts`, which restored Prettier/ESLint/TypeScript parsing. (GPT-5.5 via GitHub Copilot coding agent)
