+++ docs/ai-logs/2026-06-22-issue-762-compound-editor-factory.md
@@ -1,0 +1,79 @@
+# 2026-06-22 — Compound editor: shared state factory (#762)
+
+## Session Narrative
+
+### Prompt 1: Work on issue #762, make a PR when ready
+
+**AI response**: Issue #762 (part of the #761 architecture refactor) asked to
+split the 1,258-LOC `compound-editor-modal.svelte` by lifting its inline editor
+logic into a rune-based state factory so the desktop modal and the mobile sheet
+become thin views over one controller. Structure-only, no behaviour change.
+
+Findings before implementing:
+
+- The `compound-editor/types.ts` `EditorController` contract and
+  `compound-editor/mobile-sheet.svelte` already existed and the mobile sheet
+  was **already** a thin view over `EditorController`. The duplication the
+  issue described had been partly resolved earlier — but the controller object
+  was still defined **inline** in the desktop modal, so the logic had never
+  been extracted to a reusable module.
+- The desktop modal held all 13 `$state`/`$derived` declarations and ~24
+  functions inline, plus the full desktop template, picker dialog, and the
+  preset/formula confirm dialogs.
+
+Implementation:
+
+1. **New `state/compound-editor.svelte.ts`** — `createCompoundEditorState({ onSave })`
+   returns a `CompoundEditorState` (a desktop superset of the shared
+   `EditorController`). It owns every piece of editor state and behaviour,
+   ported verbatim from the modal: form data, element/weight texts, mode,
+   validation (`errors`/`visibleErrors`), `canSave`/`saveBlockReason`,
+   duplicate-Z (`duplicateBanner`/`editDuplicatePrompt`), mass percents,
+   display elements, paste/preset application + confirm state, URL-recovery
+   amber highlighting (`urlFailedFields`/`isUrlAmber`/`fieldBorderClass`), and
+   a `load()` lifecycle method that replaces the modal's open/compound/prefill
+   `$effect` body. Pure helpers (`deriveMassPercents`, `rescaleTo100`,
+   `computeAtomCounts`, …) are still called, not duplicated.
+2. **New `compound-editor/desktop-sheet.svelte`** — the desktop template moved
+   here verbatim, bound to the controller. Keeps only genuinely view-local
+   state: `showDeleteConfirm`, picker `pickerMode`/`pickerEditIndex`, and
+   `confirmRemoveIndex`. Hosts the picker dialog and the preset/formula confirm
+   dialogs (effectively desktop-only — on phones the quick-start panel only
+   shows for an empty composition, so those confirmations never trigger).
+3. **`compound-editor-modal.svelte`** reduced to a 111-LOC orchestrator:
+   creates the controller, runs `editor.load()` from a `$effect`, owns the
+   `isMobile`/`prefersReducedMotion` media queries, and renders either
+   `MobileSheet` or `DesktopSheet` inside the bits-ui Dialog shell.
+4. Behaviour-preserving simplification: the desktop "add by symbol" handler now
+   calls `editor.addElementBySymbol()` (deduped with the mobile path); the
+   weight inputs use `value` + `setWeightText` like mobile instead of a
+   `bind:value` + ad-hoc `compositionTouched` write.
+
+Verification: `pnpm run check` (0 errors), `pnpm lint`, `pnpm format:check`
+clean; full Vitest suite **1772 passed / 3 skipped** including the unchanged
+`compound-editor-modal.test.ts` component test and the new factory unit test.
+
+## Tasks
+
+### Extract shared compound-editor state factory
+
+- **Status**: completed
+- **Stage**: architecture refactor (#761 → #762)
+- **Files changed**:
+  - NEW `src/lib/state/compound-editor.svelte.ts`
+  - NEW `src/lib/components/compound-editor/desktop-sheet.svelte`
+  - `src/lib/components/compound-editor-modal.svelte` (1,258 → 111 LOC)
+  - `src/lib/components/compound-editor/types.ts` (doc comment)
+  - NEW `src/tests/unit/compound-editor-state.test.ts`
+  - `docs/03-architecture.md` (§8 + state/component listings)
+  - `CHANGELOG-AI.md`, this log
+- **Decision**: Kept `EditorController` as the shared (mobile) contract and had
+  the factory return a `CompoundEditorState` superset for the desktop extras
+  (`load`, confirm-dialog state, URL-amber helpers) rather than widening the
+  shared interface with desktop-only members.
+- **Decision**: `load()` replaces the reactive reset effect; the modal calls it
+  inside `untrack` from a `$effect` that tracks `open`/`compound`/`prefill`/
+  `initialWarning`, matching the previous semantics.
+- **Issue**: none outstanding. E2E specs (`custom-compounds.spec.ts`,
+  `compound-editor-mobile.spec.ts`) require built WASM and were not run locally;
+  they exercise only public behaviour, which is unchanged.
