+++ docs/ai-logs/README.md
@@ -19,6 +19,7 @@ summary table of all sessions.
 
 | File | Date | Topic |
 | ------------------------------------------------------------------ | -------------------------------------------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [2026-06-22-issue-762-compound-editor-factory.md](2026-06-22-issue-762-compound-editor-factory.md) | 22 June 2026 | **Split compound-editor-modal into a shared state factory (#762)** (Claude Opus 4.8 via Claude Code): extracted the inline editor logic into `src/lib/state/compound-editor.svelte.ts` (`createCompoundEditorState`) and the desktop template into `compound-editor/desktop-sheet.svelte`; the modal is now a 111-LOC orchestrator and the desktop + mobile sheets are thin views over one controller (no duplicated logic). Added direct factory unit tests; component + e2e behaviour unchanged. Structure-only. |
 | [2026-06-21-contextual-help.md](2026-06-21-contextual-help.md) | 21 June 2026 | **Contextual help foundation + "Program / data source" explainer (#769)** (Claude Opus 4.8 via Claude Code): new accessible `HelpHint` ⓘ component + glossary-sourced `help-text.ts` registry (WCAG 1.4.13); `PROGRAM_HELP` map; Program-panel concept explainer + per-program hints in `program-tab.svelte`; new glossary term "Program (Data Source)" and a user-guide "Choosing a program" section; `contextual-help.md` spec. Issues #770/#771 cover the remaining quantity/unit (with CSDA-range parity) and advanced/workflow hints. |
 | [2026-06-21-issue-767-compound-editor-ux.md](2026-06-21-issue-767-compound-editor-ux.md) | 21 June 2026 | **Compound Editor UX hardening (#767)** (Claude Opus 4.8 via Claude Code): deferred validation (errors show on blur / Save attempt via a `visibleErrors` derived), Save stays clickable with a visible `role="alert"` block reason instead of a tooltip, `aria-invalid`/`aria-describedby` on field inputs, and emoji icons replaced with lucide. |
 | [2026-06-21-issue-763-orchestrator-decompose.md](2026-06-21-issue-763-orchestrator-decompose.md) | 21 June 2026 | **Decompose calculator-page orchestrator (#763)** (Claude Opus 4.8 via Claude Code): extracted the shared-compound-from-URL flow (`sharedUrl*`/recovery-editor state + six methods) into `src/lib/state/shared-compound-from-url.svelte.ts` with a focused test; orchestrator delegates via getters/methods, keeping `calculator/+page.svelte` usage unchanged. Structure-only. |
